Added new radial left hook auto
SmoothLeftHookApproach should have the both positioned 12.5 inches from the boiler flush against the alliance stations. It will slowly accelerate in a curve towards the hook in a given time, in seconds, that can be altered to go faster or slower. Regardless of the time, the robot will move a constant distance.